sdk/python/pyproject.toml
[project]
name = "hx-loli-img"
version = "0.1.0"
description = "HX LoLi Img Python SDK - 上传图片到 CF R2"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27",
    "pillow>=10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
